store mSystemMemoryEnd to Memory(>0x50000000)

+++ Tiny6410Pkg/PrePi/PrePi.h
@@ -29,7 +29,7 @@
 
 #define SerialPrint(txt)  SerialPortWrite (txt, AsciiStrLen(txt)+1);
 
-extern UINT64 mSystemMemoryEnd;
+extern UINT32 mSystemMemoryEnd;
 
 RETURN_STATUS
 EFIAPI
